backend/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
idna==3.11
imbalanced-learn==0.14.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lightgbm==4.6.0
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
PyJWT==2.10.1
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==3.1.2
